Implement `initial_diff` function
