+++ src/vcs/git/repo.rs
@@ -16,7 +16,6 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    diff,
     diff::*,
     file_system,
     vcs,
@@ -178,93 +177,14 @@ impl<'a> RepositoryRef<'a> {
 
     /// Get the [`Diff`] between two commits.
     pub fn diff(&self, from: Oid, to: Oid) -> Result<Diff, Error> {
-        use git2::{Delta, Patch};
-
-        let mut diff = Diff::new();
         let git_diff = self.diff_commits(None, Some(from), to)?;
+        Diff::try_from(git_diff)
+    }
 
-        for (idx, delta) in git_diff.deltas().enumerate() {
-            match delta.status() {
-                Delta::Added => {
-                    let diff_file = delta.new_file();
-                    let path = diff_file.path().ok_or(diff::git::Error::PathUnavailable)?;
-                    let path = file_system::Path::try_from(path.to_path_buf())?;
-
-                    diff.add_created_file(path);
-                },
-                Delta::Deleted => {
-                    let diff_file = delta.old_file();
-                    let path = diff_file.path().ok_or(diff::git::Error::PathUnavailable)?;
-                    let path = file_system::Path::try_from(path.to_path_buf())?;
-
-                    diff.add_deleted_file(path);
-                },
-                Delta::Modified => {
-                    let diff_file = delta.new_file();
-                    let path = diff_file.path().ok_or(diff::git::Error::PathUnavailable)?;
-                    let path = file_system::Path::try_from(path.to_path_buf())?;
-
-                    let patch = Patch::from_diff(&git_diff, idx)?;
-
-                    if let Some(patch) = patch {
-                        let mut hunks: Vec<Hunk> = Vec::new();
-
-                        for h in 0..patch.num_hunks() {
-                            let (hunk, hunk_lines) = patch.hunk(h)?;
-                            let header = Line(hunk.header().to_owned());
-                            let mut lines: Vec<LineDiff> = Vec::new();
-
-                            for l in 0..hunk_lines {
-                                let line = patch.line_in_hunk(h, l)?;
-                                let line = LineDiff::try_from(line)?;
-                                lines.push(line);
-                            }
-                            hunks.push(Hunk { header, lines });
-                        }
-                        diff.add_modified_file(path, hunks);
-                    } else if diff_file.is_binary() {
-                        diff.add_modified_binary_file(path);
-                    } else {
-                        return Err(diff::git::Error::PatchUnavailable(path).into());
-                    }
-                },
-                Delta::Renamed => {
-                    let old = delta
-                        .old_file()
-                        .path()
-                        .ok_or(diff::git::Error::PathUnavailable)?;
-                    let new = delta
-                        .new_file()
-                        .path()
-                        .ok_or(diff::git::Error::PathUnavailable)?;
-
-                    let old_path = file_system::Path::try_from(old.to_path_buf())?;
-                    let new_path = file_system::Path::try_from(new.to_path_buf())?;
-
-                    diff.add_moved_file(old_path, new_path);
-                },
-                Delta::Copied => {
-                    let old = delta
-                        .old_file()
-                        .path()
-                        .ok_or(diff::git::Error::PathUnavailable)?;
-                    let new = delta
-                        .new_file()
-                        .path()
-                        .ok_or(diff::git::Error::PathUnavailable)?;
-
-                    let old_path = file_system::Path::try_from(old.to_path_buf())?;
-                    let new_path = file_system::Path::try_from(new.to_path_buf())?;
-
-                    diff.add_copied_file(old_path, new_path);
-                },
-                status => {
-                    return Err(diff::git::Error::DeltaUnhandled(status).into());
-                },
-            }
-        }
-
-        Ok(diff)
+    /// Get the [`Diff`] of a commit with no parents.
+    pub fn initial_diff(&self, oid: Oid) -> Result<Diff, Error> {
+        let git_diff = self.diff_commits(None, None, oid)?;
+        Diff::try_from(git_diff)
     }
 
     pub(super) fn switch_namespace(&self, namespace: &str) -> Result<(), Error> {
